fix: Tests for governance plugin
