+++ packages/dai-plugin-governance/test/ChiefService.test.js
@@ -26,10 +26,10 @@ beforeAll(async () => {
   maker.useAccount('owner');
 });
 
-afterAll(async done => {
+afterAll(async () => {
   if (global.useOldChain) {
     await restoreSnapshotOriginal(global.snapshotId);
-    done();
+    return;
   } else {
     global.client.restoreSnapshot(global.testchainId, global.defaultSnapshotId);
     await sleep(15000);
@@ -37,7 +37,7 @@ afterAll(async done => {
     await global.client.delete(global.testchainId);
     await sleep(15000);
 
-    done();
+    return;
   }
 });
