+++ packages/dai-plugin-governance/test/EsmService.test.js
@@ -17,10 +17,10 @@ beforeAll(async () => {
   await setUpAllowance(maker, esmService._esmContract().address);
 });
 
-afterAll(async done => {
+afterAll(async () => {
   if (global.useOldChain) {
     await restoreSnapshotOriginal(global.snapshotId);
-    done();
+    return;
   } else {
     global.client.restoreSnapshot(global.testchainId, global.defaultSnapshotId);
     await sleep(15000);
@@ -28,7 +28,7 @@ afterAll(async done => {
     await global.client.delete(global.testchainId);
     await sleep(15000);
 
-    done();
+    return;
   }
 });
