+++ packages/dai-plugin-governance/test/GovPollingService.test.js
@@ -42,16 +42,16 @@ beforeAll(async () => {
   maker.useAccount('owner');
 });
 
-afterAll(async done => {
+afterAll(async () => {
   if (global.useOldChain) {
     await restoreSnapshotOriginal(global.snapshotId);
-    done();
+    return;
   } else {
     global.client.restoreSnapshot(global.testchainId, global.defaultSnapshotId);
     await sleep(15000);
     await global.client.delete(global.testchainId);
     await sleep(15000);
-    done();
+    return;
   }
 });
