+++ packages/dai-plugin-governance/test/SpellService.test.js
@@ -6,10 +6,10 @@ import {
 
 let maker, spellService;
 
-afterAll(async done => {
+afterAll(async () => {
   if (global.useOldChain) {
     await restoreSnapshotOriginal(global.snapshotId);
-    done();
+    return;
   } else {
     global.client.restoreSnapshot(global.testchainId, global.defaultSnapshotId);
     await sleep(15000);
@@ -17,7 +17,7 @@ afterAll(async done => {
     await global.client.delete(global.testchainId);
     await sleep(15000);
 
-    done();
+    return;
   }
 });
